more cleanup/code moving
